fix get_model_response_iterator
